Added the Projects Folder
The Projects Folder shall contain the special project portions found throughout the course. These are longer lessons that combine many things learned throughout the course so far. I shall try to add more of my own code into it, to alter the game mode exactly the way I want it, as well as to experiment and learn more about spigot and all the plethora of things you can do with plugins.