Private Messaging System Project
This project was for making a command that lets a player message another player privately using the command /message <player> <message>. If the message is sent properly to the target player, then the UUID of both players is put into a HashMap named recentMessages. This is used to tell if the players have interacted at least once, and will let the target player to use the reply command. The target player can reply to the message without entering the initial player's username, using the command /reply <message>.