Finish up the More Advanced API Section
Cooldown.java - Uses a google Cache with UUID, Long to act as a cooldown hashmap. Using this special hashmap made from CacheBuilder, you can add a custom cooldown to a command, as shown in here.
Display Entities - Done originally in the main class, shows how to implement the three different display entities added in 1.19.4 into the API. These three are Text Displays, Block Displays, and Item Displays.
SpawnHologramCommand.java - Implements a command that lets a user create a hologram. You need to first us an Armor stand and set it to invisible and invulnerable, set gravity to false, set custom name visible to true, and set custom name to the lines of the String array[] created earlier. This will give the invisible armorstand a visible text display of the lines in the array ingame. This lets you create text holograms.
PermsCommand - Shows to add and remove perms from a player using the PermissionAttachment class. Need to use HashMap<UUID, PermissionAttachment).
SidebarListener.java - Shows how to implement a sidebar for the player that keeps track of mobs slain.

Nametags:
Rank.java - Enum class that holds the ranks that will be used in the nametags.
NametagListener.java - When player joins the server, it will use the NametagManager to set the nametags for the player so they can see ranks, and add the player to the teams of other players so they can see the ranks too.
NametagManager.java - Creates the scoreboard and teams that controls how the rank is seen in the scoreboard by players.