Rank System Project
This project was for implementing an entire rank system that uses a wide variety of features to create in-game ranks for players. An enum class is used for the ranks, called Ranks. The RankManager is used to get and set ranks, as well as perms. It also creates and reads the yml file where the player's uuid and rank is saved.  The NametagManager utilizes the rank information from RankManager to make a scoreboard for every player in the server. Teams are created for each rank, and all players are added to it so they can see each other's ranks. RankListener exists so that player's get a rank when they join the server, and it gets removed when they leave. There is also an event for getting rid of any messages they send, and broadcasting their original message with their rank added on to it.